refactor: reorganize UUID and ImageProcessing utilities for better structure
