+++ src/lib/utils/uuid.ts
@@ -1,0 +1,5 @@
+export const empty = '00000000-0000-0000-0000-000000000000';
+
+export function isEmpty(uuid: string): boolean {
+    return uuid === empty;
+}
